feat(cc): complete rework of terraform cookiecutter to be generalized and opinionated
This template should be simpler to consume as it takes away a few
options and generates projects in a more consistent state.

This decision was made to promote project scaffolding in what I view to
be the "correct" way... as opposed to allowing users to opt out of
controls they really should enforce, such as scm, git hooks, etc.